% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_metadata_seurat.R
\name{align_metadata_seurat}
\alias{align_metadata_seurat}
\title{Align Metadata to Cell Names in Seurat Object}
\usage{
align_metadata_seurat(data, combined_counts)
}
\arguments{
\item{data}{The Seurat object.}

\item{combined_counts}{The combined counts matrix.}
}
\value{
A metadata DataFrame aligned with the counts matrix.
}
\description{
Ensures that the metadata is correctly aligned with the cell names in the counts matrix.
}
