% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_alt_experiments_sce.R
\name{attach_alt_experiments_sce}
\alias{attach_alt_experiments_sce}
\title{Attach altExps to SingleCellExperiment Object}
\usage{
attach_alt_experiments_sce(data, sce, altExp_names)
}
\arguments{
\item{data}{The Seurat object.}

\item{sce}{The SingleCellExperiment object.}

\item{altExp_names}{A vector of altExp names.}
}
\value{
The SingleCellExperiment object with attached altExps.
}
\description{
Attaches alternative experiments to the SingleCellExperiment object.
}
