% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_commands.R
\name{convert_commands}
\alias{convert_commands}
\title{Convert SeuratCommand objects into a serializable format}
\usage{
convert_commands(commands_list)
}
\arguments{
\item{commands_list}{A list of SeuratCommand objects from Seurat metadata.}
}
\value{
A list of simplified, serializable command information.
}
\description{
This function takes a list of SeuratCommand objects and converts them into
JSON-serializable lists, preserving command name, parameters, timestamps,
and the assay used.
}
