% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_combined_assay.R
\name{create_combined_assay}
\alias{create_combined_assay}
\title{Create Combined Assay in Seurat Object}
\usage{
create_combined_assay(data, combined_counts, condition_labels)
}
\arguments{
\item{data}{The Seurat object.}

\item{combined_counts}{The combined counts matrix.}

\item{condition_labels}{Labels indicating the origin of each cell.}
}
\value{
The updated Seurat object.
}
\description{
Creates a new assay in the Seurat object using the combined counts matrix.
}
