% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_csparse_matrix.R
\name{ensure_csparse_matrix}
\alias{ensure_csparse_matrix}
\title{Ensure Sparse Matrix is in a Format Compatible with AnnData}
\usage{
ensure_csparse_matrix(mat)
}
\arguments{
\item{mat}{A matrix or Matrix object}
}
\value{
The original matrix if already in a compatible format, or a converted
        matrix if conversion was needed
}
\description{
This function checks if a sparse matrix is in a format compatible with AnnData 
(CsparseMatrix format, e.g., dgCMatrix) and converts it if necessary.
}
