% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_counts_matrix.R
\name{extract_counts_matrix}
\alias{extract_counts_matrix}
\title{Extract Counts Matrix from Seurat Assay}
\usage{
extract_counts_matrix(assay_object)
}
\arguments{
\item{assay_object}{The default assay object from the Seurat object.}
}
\value{
A counts matrix.
}
\description{
Retrieves the counts matrix from the default assay of the Seurat object.
}
