% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extract Data from SingleCellExperiment}
\usage{
extract_data(data_fun, data_type, sce)
}
\arguments{
\item{data_fun}{A function to extract data (e.g., colData, rowData).}

\item{data_type}{The type of data being extracted (e.g., "obs/colData", "var/rowData").}

\item{sce}{A SingleCellExperiment object.}
}
\value{
A list containing the extracted data and internal columns.
}
\description{
This function extracts data from the SingleCellExperiment object with careful error handling.
}
