% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pairs.R
\name{extract_pairs}
\alias{extract_pairs}
\title{Extract Pairwise Data from SingleCellExperiment}
\usage{
extract_pairs(pairs_function, pair_type, sce)
}
\arguments{
\item{pairs_function}{A function to extract pairwise data (e.g., colPairs or rowPairs).}

\item{pair_type}{A string indicating whether the data is colPairing or rowPairing.}

\item{sce}{A SingleCellExperiment object from which to extract the pairwise data.}
}
\value{
A list of pairwise data as sparse matrices.
}
\description{
This function extracts pairwise data (colPairs or rowPairs) from a SingleCellExperiment
object and returns it as a list of sparse matrices.
}
