% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_alt_exps_seurat.R
\name{identify_alt_exps_seurat}
\alias{identify_alt_exps_seurat}
\title{Identify Assays to Move to altExps in Seurat Object}
\usage{
identify_alt_exps_seurat(data)
}
\arguments{
\item{data}{The Seurat object.}
}
\value{
A vector of assay names to be moved to `altExps`.
}
\description{
Identifies assays with mismatching cells or features and prepares them to be moved to `altExps`.
}
