% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_experiments.R
\name{process_alt_experiments}
\alias{process_alt_experiments}
\title{Process Alternative Experiments}
\usage{
process_alt_experiments(sce, assayName, useAltExp)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{assayName}{The name of the assay to use as the main data matrix in AnnData.}

\item{useAltExp}{Logical indicating whether to process and include alternative experiments (altExps).}
}
\value{
A list containing the updated SingleCellExperiment object and the altExps data.
}
\description{
This function processes alternative experiments (altExps) in the SingleCellExperiment object.
If altExps are found, they are processed recursively and stored in the `uns` slot of the AnnData object.
}
