% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reductions.R
\name{process_dimensional_reductions}
\alias{process_dimensional_reductions}
\title{Process Dimensional Reductions}
\usage{
process_dimensional_reductions(sce)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}
}
\value{
A list of dimensional reductions for the AnnData object.
}
\description{
This function processes the dimensional reductions in the SingleCellExperiment object.
}
