% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_layers.R
\name{process_layers}
\alias{process_layers}
\title{Process Layers in Seurat Assay}
\usage{
process_layers(assay_object)
}
\arguments{
\item{assay_object}{The Seurat Assay5 object to process.}
}
\value{
A list with split names per layer and updated assay object.
}
\description{
Combines split layers in a Seurat Assay5 object, stores them as non-split layers, and removes the split layers.
}
