% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{process_metadata_and_pairwise}
\alias{process_metadata_and_pairwise}
\title{Process Metadata and Pairwise Matrices}
\usage{
process_metadata_and_pairwise(sce, alt_exps_data, X)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{alt_exps_data}{Data of alternative experiments.}

\item{X}{The main data matrix (anndata.X) for the AnnData object.}
}
\value{
A list containing metadata and pairwise matrices.
}
\description{
This function processes the metadata and pairwise matrices in the SingleCellExperiment object.
}
