% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_assays.R
\name{process_other_assays}
\alias{process_other_assays}
\title{Process Other Assays}
\usage{
process_other_assays(sce, assayName = NULL)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{assayName}{(Optional) The name of the assay used as the main data matrix.
If not provided, all assays will be processed.}
}
\value{
A list of processed assays.
}
\description{
This function processes assays other than the main assay in the SingleCellExperiment object.
}
