% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamped_cat.R
\name{timestamped_cat}
\alias{timestamped_cat}
\title{Print Messages with Timestamp}
\usage{
timestamped_cat(...)
}
\arguments{
\item{...}{The messages to print.}
}
\value{
None. The function prints the messages to the console.
}
\description{
This function prints messages with a timestamp.
}
\examples{
timestamped_cat("This is a message.")
}
