% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_seurat_object.R
\name{update_seurat_object}
\alias{update_seurat_object}
\title{Update Seurat Object if Necessary}
\usage{
update_seurat_object(data)
}
\arguments{
\item{data}{A Seurat object.}
}
\value{
An updated Seurat object.
}
\description{
Checks if the Seurat object is from an old version (v2) and updates it to the latest format.
}
