% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_convert.R
\name{cli_convert}
\alias{cli_convert}
\title{Command Line Interface for convert2anndata}
\usage{
cli_convert()
}
\description{
This function serves as a command line interface for the convert2anndata package.
It parses command line arguments and calls the appropriate functions to convert
a SingleCellExperiment or Seurat object to an AnnData object.
}
