% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preserve_reductions_seurat.R
\name{preserve_reductions_seurat}
\alias{preserve_reductions_seurat}
\title{Preserve Dimensional Reductions from Seurat Object}
\usage{
preserve_reductions_seurat(data, sce)
}
\arguments{
\item{data}{The Seurat object.}

\item{sce}{The SingleCellExperiment object.}
}
\value{
The SingleCellExperiment object with updated dimensional reductions.
}
\description{
Updates and transfers dimensional reductions from the Seurat object to the SingleCellExperiment object.
}
