% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_metadata_and_pairwise.R
\name{process_metadata_and_pairwise}
\alias{process_metadata_and_pairwise}
\title{Process Metadata and Pairwise Matrices}
\usage{
process_metadata_and_pairwise(sce, alt_exps_data, X, obsm = list())
}
\arguments{
\item{sce}{A `SingleCellExperiment` object containing the data to be processed.}

\item{alt_exps_data}{A list of alternative experiment data to be added to the metadata.}

\item{X}{The main data matrix (`anndata.X`) for the AnnData object. This is used to validate
dimensions of metadata and pairwise matrices.}

\item{obsm}{A list of observation matrices, which can be extended during processing.}
}
\value{
A list containing:
  \item{uns}{A list of remaining metadata that was not categorized into other slots.}
  \item{obsp}{A list of pairwise observation matrices.}
  \item{varp}{A list of pairwise variable matrices.}
  \item{obsm}{A list of observation matrices.}
  \item{varm}{A list of variable matrices.}
}
\description{
This function processes the metadata and pairwise matrices in a `SingleCellExperiment` object.
It organizes metadata into distinct categories (`uns`, `obsp`, `obsm`, `varp`, `varm`) based
on their dimensions and relationship to the main data matrix (`X`). Metadata matrices matching
observation or variable dimensions are moved into the corresponding slots, while remaining
metadata is retained in `uns`. Pairwise matrices for observations and variables are identified
and organized into `obsp` and `varp`, respectively.
}
