% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_assays.R
\name{process_other_assays}
\alias{process_other_assays}
\title{Process Other Assays}
\usage{
process_other_assays(sce, assayName)
}
\arguments{
\item{sce}{A SingleCellExperiment object.}

\item{assayName}{The name of the assay used as the main data matrix.}
}
\value{
A list of other assays for the AnnData object.
}
\description{
This function processes assays other than the main assay in the SingleCellExperiment object.
}
